% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astd.r
\name{astd}
\alias{astd}
\title{Adjusted Standard Deviation}
\usage{
astd(...)
}
\arguments{
\item{series}{numeric or ts object}
}
\value{
numeric or ts object
}
\description{
This function calculates the adjusted standard deviation from the last 12 months
of a time series.
}
\examples{
\dontrun{
astd(AirPassengers)
}
}
\author{
Obryan Poyser
}
