% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_volval.r
\name{class_volval}
\alias{class_volval}
\title{Returns the classification of a GMID according to its value or volume}
\usage{
class_volval(series)
}
\arguments{
\item{series}{numeric vector}
}
\value{
string
}
\description{
This is simple function that estimates the classification of an item given that its weight in the entire sample.
The classification is created as follow: $x <= .80 ~ A, x <= .95 ~ B, otherwise ~ C$.
}
\examples{
\dontrun{
class_volval(AirPassengers)
}
}
\author{
Obryan Poyser
}
