% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvts.R
\name{cvts}
\alias{cvts}
\title{Cross-validation for time series data}
\usage{
cvts(ts_data, cv_size, lag)
}
\arguments{
\item{ts_data}{tibble, dataframe or matrix. It is assumed that it is ordered by date.}

\item{cv_size}{numeric. Cross validation size.}

\item{lag}{numeric. How many periods ahead to forecast.}
}
\value{

}
\description{
Cross-validation for time series data
}
\examples{
\dontrun{
cvts()
}
}
