% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_reg.R
\name{empty_reg}
\alias{empty_reg}
\title{Generate regressors tables' template}
\usage{
empty_reg()
}
\value{
two CSV files on users' desktop
}
\description{
This function creates two CSV files on users' desktop. One is mean to
include regressors' values and other for regressors' names,
for Regressors and Edit Regressors Values respectively.
}
\examples{
\dontrun{
empty_reg()
}
}
\author{
Obryan Poyser
}
