% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_reg.r
\name{harvest_reg}
\alias{harvest_reg}
\title{Harvest Kinaxis regressors}
\usage{
harvest_reg(reg_list_path, reg_values_path, dir_path)
}
\arguments{
\item{reg_list_path}{String. Path where the **"Regressor.tab"** file is located}

\item{reg_values_path}{String. Path where the **"Edit Regressor Values.tab"** file is located}

\item{dir_path}{String. Directory where all the CSV files are located}
}
\value{
reg_list and reg_values for regressors names and regressors values. Ultimately two tab files "Regressors.tab" and "Edit Regressors Values.tab"
}
\description{
This functions perform several loading and sanity checks to regressor lists. The process starts
by attaching to the current environment the most updated version of **"Regressor.tab"** and
**"Edit Regressor Values.tab"** tables from Kinaxis. Then attachs the new regressors to be added by each user,
included in a single directory with filenames following the format <userid>_reg_list.csv and
<userid>_reg_values.csv. Finally, the function appends zero values when neccessary, and performs sanity
checks to prevent importing errors in Kinaxis.
}
\examples{
\dontrun{
harvest_reg()
}
}
