% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_descriptor.R, R/load_gbu.R
\name{load_gbu}
\alias{load_gbu}
\title{Get Global Business Unit data}
\usage{
load_gbu(gbu_path, on_globalenv = F)

load_gbu(gbu_path, on_globalenv = F)
}
\arguments{
\item{gbu_path}{string. Has the default directory}

\item{on_globalenv}{logical. Defines is the dataframe should be loaded in the global environment or not.}

\item{db}{string. Defines which database to load. For now only two are available.
#' Options: *GBU*, *LS*.}
}
\value{
tibble

tibble
}
\description{
Analytics' teams manages a database of Global Business Unit that links Location, GMID.
There is a default directory located in ~Active/Specific.GBU.Rdata.

Analytics' teams manages a database of Global Business Unit that links Location, GMID.
There is a default directory located in ~Active/Specific.GBU.Rdata.
}
\examples{
\dontrun{
load_gbu()
}
\dontrun{
load_gbu(on_globalenv = T)
}
}
\author{
Obryan Poyser

Obryan Poyser
}
