% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lifesavers.r
\name{load_lifesavers}
\alias{load_lifesavers}
\title{Get life savers data}
\usage{
load_lifesavers(lifesavers_path)
}
\arguments{
\item{lifesavers_path}{string. Has the default directory}
}
\value{
tibble
}
\description{
Get life savers data
}
\examples{
\dontrun{
load_lifesavers()
}
}
\author{
Obryan Poyser
}
