% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_reg_inputs.R
\name{load_reg_inputs}
\alias{load_reg_inputs}
\title{Loads regressors' inputs}
\usage{
load_reg_inputs(reg_list_path, reg_values_path)
}
\arguments{
\item{reg_list_path}{String. Path where the Regressors' names are located}

\item{reg_values_path}{String. Path where the Regressors' values are located}
}
\value{
reg_list and reg_values for regressors names and regressors values
}
\description{
Loading input .tab files from Kinaxis tables Regressors and Edit Regressors Values' worksheets
}
\examples{
load_reg_inputs()
}
