% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_method.R
\name{outlier_method}
\alias{outlier_method}
\title{Performs outlier method to Kinaxis data}
\usage{
outlier_method(
  time_series,
  outlier_method,
  data_rule,
  mov_avg_n,
  threshold,
  causal_factor = T
)
}
\arguments{
\item{time_series}{ts object. The signal were to applied the outlier method.}

\item{outlier_method}{String. Defines the type of outlier method to be applied.
Options: *Standard Deviation*, *Iglewicz Hoaglin Method*, *Winsorizing* and *IQR*.}

\item{data_rule}{String. Defines the classification of the data according to the number of observations.
Options: *Historical*, *Moving Average Error* and *Rstl Error*.}

\item{mov_avg_n}{Integer. Defines how many periods to consider to smooth the series.}

\item{threshold}{Double. Defines the interval cut-off the extreme values. Between $0$ and $1$.}

\item{causal_factor}{Logical. Defines if the series have regressors or not.}
}
\value{
ts object.
}
\description{
Performs outlier method to Kinaxis data
}
\examples{
\dontrun{
outlier_method()
}
}
\author{
Sze Gee
}
