% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata2obj.R
\name{rdata2obj}
\alias{rdata2obj}
\title{RData to object}
\usage{
rdata2obj(path)
}
\arguments{
\item{path}{String. Path to RData file}
}
\value{

}
\description{
This simple function receives a RData and passes as object to be assigned
}
\examples{
\dontrun{
rdata2object("path")
}
}
\author{
Obryan Poyser
}
