% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFtools.R
\docType{package}
\name{SFtools-package}
\alias{SFtools}
\alias{SFtools-package}
\title{SFtools: Space Filling Based Tools for Data Mining}
\description{
Contains space filling based tools for
    machine learning and data mining. Some functions offer
    several computational techniques and deal with the out of
    memory for large big data by using the ff package.
}
\references{
M. Laib, M. Kanevski, 
\href{https://www.elen.ucl.ac.be/Proceedings/esann/esannpdf/es2018-57.pdf}{A novel 
filter algorithm for unsupervised feature selection based on a space filling measure}. 
Proceedings of the 26rd European Symposium on Artificial Neural Networks, Computational 
Intelligence and Machine Learning (ESANN), pp. 485-490, Bruges (Belgium), 2018.

M. Laib and M. Kanevski, A new algorithm for redundancy minimisation in 
geo-environmental data, 2019.
\href{https://www.sciencedirect.com/science/article/pii/S0098300418310975}{Computers & 
Geosciences, 133 104328}.

J. A. Royle, D. Nychka, An algorithm for the construction of spatial
coverage designs with implementation in Splus, Computers and Geosciences
24 (1997) p. 479–488.

J. Franco, Planification d’expériences numériques en phase exploratoire
pour la simulation des phénomènes complexes, Thesis (2008) 282.

D. Dupuy, C. Helbert, J. Franco (2015). DiceDesign
and DiceEval: Two R Packages for Design and Analysis of Computer
Experiments. Journal of Statistical Software, 65(11), 1-38.
\href{http://www.jstatsoft.org/v65/i11/}{Jstatsoft}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlaib.github.io/}
  \item Report bugs at \url{https://github.com/mlaib/SFtools/issues}
}

}
\author{
Mohamed Laib \email{Mohamed.Laib@gmail.com}  and

 Mikhail Kanevski \email{Mikhail.Kanevski@unil.ch},

Maintainer: Mohamed Laib \email{laib.med@gmail.com}
}
