% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimData.R
\name{SimData}
\alias{SimData}
\title{Simulated data set}
\usage{
SimData(n=1000)
}
\arguments{
\item{n}{Number of generated data points (by default: \code{n=1000}).}
}
\value{
A \code{data.frame} of simulated data set, with \eqn{7} features
(\eqn{4} of them are redundants)
}
\description{
Generates a simulated data set
}
\examples{
Sim_Data<-SimData(n=1000)
plot(Sim_Data$x1,Sim_Data$x2)

\dontrun{

#### Visualisation of the data set (3D) ####
require(rgl)
require(colorRamps)

c <- cut(Sim_Data$z,breaks=100)
cols <- matlab.like(100)[as.numeric(c)]
plot3d(Sim_Data$x1,Sim_Data$x2,Sim_Data$z,radius=0.01, col=cols,
type="s",xlab="x1",ylab="x2",zlab="z",box=F)
grid3d(c("x","y","z"),col="black",lwd=1)

}
}
