% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_twelve.R
\name{sum_twelve}
\alias{sum_twelve}
\title{Series sum of the last 12 months}
\usage{
sum_twelve(...)
}
\arguments{
\item{series}{numeric of ts object}
}
\value{
numeric
}
\description{
Series sum of the last 12 months
}
\examples{
sum_twelve(AirPassengers)
}
\author{
Obryan Poyser
}
