% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UfsCov_ff.R
\name{UfsCov_ff}
\alias{UfsCov_ff}
\title{UfsCov for unsupervised features selection}
\usage{
UfsCov_ff(data, blocks=2)
}
\arguments{
\item{data}{Data of class: \code{matrix} or \code{data.frame}.}

\item{blocks}{Number of splits to facilitate the computation of the
distance matrix (by default: blocks=2).}
}
\value{
A list of two elements:
 \itemize{
  \item \code{CovD} a vector containing the coverage measure of
  each step of the SFS.
  \item \code{IdR} a vector containing the added variables during
  the selection procedure.
  }
}
\description{
Applies the UfsCov algorithm based on the space filling concept,
by using a sequatial forward search (for memory efficient storage
of large data on disk and fast access).
}
\note{
This function is still under developement.
}
\examples{

\dontrun{
#### Infinity dataset ####
N <- 1000
dat<-Infinity(N)
Results<- UfsCov_ff(dat)

cou<-colnames(dat)
nom<-cou[Results[[2]]]
par(mfrow=c(1,1), mar=c(5,5,2,2))
names(Results[[1]])<-cou[Results[[2]]]
plot(Results[[1]] ,pch=16,cex=1,col="blue", axes = FALSE,
xlab = "Added Features", ylab = "Coverage measure")
lines(Results[[1]] ,cex=2,col="blue")
grid(lwd=1.5,col="gray" )
box()
axis(2)
axis(1,1:length(nom),nom)
which.min(Results[[1]])

#### Butterfly dataset ####

require(IDmining)
N <- 1000
raw_dat <- Butterfly(N)
dat<-raw_dat[,-9]

Results<- UfsCov_ff(dat)

}
}
\references{
M. Laib and M. Kanevski (2017). Unsupervised Feature Selection Based on Space
Filling Concept, \href{https://arxiv.org/abs/1706.08894}{arXiv:1706.08894}.
}
\author{
Mohamed Laib \email{Mohamed.Laib@unil.ch}
}
