% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorize.r
\name{winsorize}
\alias{winsorize}
\title{Winsorizing time series data}
\usage{
winsorize(time.series, frequency = 12)
}
\arguments{
\item{time.series}{either a ts object or numeric vector}
}
\value{
ts object or numeric vector
}
\description{
This function performs a signal winsorization. It takes a ts object or a numeric vector and returns cleansed values.
For series with more than 24 observation the function winsorizes the remainder of a STL decomposition.
}
\examples{
winsorize(AirPassengers)
}
\author{
Jillian Teo
}
