% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acov.r
\name{acov}
\alias{acov}
\title{Adjusted Coefficient of Variation}
\usage{
acov(...)
}
\arguments{
\item{series}{numeric or ts object}
}
\value{
numeric or ts object
}
\description{
This function calculates the adjusted coefficient of variation from the last 12 months
of a time series.
}
\examples{
\dontrun{
acov(AirPassengers)
}
}
\author{
Obryan Poyser
}
