% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_reg.r
\name{harvest_reg}
\alias{harvest_reg}
\title{Harvest team members regressors values}
\usage{
harvest_reg(dir_path)
}
\arguments{
\item{dir_path}{String. Directory where all the CSV files are located.}
}
\value{
Depends on users' selections. Ultimately two tab files "Regressors.tab" and "Edit Regressors Values.tab"
}
\description{
This functions perform several task, first it loads the files into the environment then
brings the possibility to do sanity checks to prevent importing error in Kinaxis
}
\examples{
harvest_reg()
}
