% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gbu.R
\name{load_gbu}
\alias{load_gbu}
\title{Get GBU data}
\usage{
load_gbu(gbu_path)
}
\arguments{
\item{gbu_path}{string. Has the default directory}
}
\value{
tibble
}
\description{
Get GBU data
}
\examples{
\dontrun{
load_gbu()
}
}
\author{
Obryan Poyser
}
