% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_knx.R
\name{read_knx}
\alias{read_knx}
\title{Maps the file name and assigns its respective function}
\usage{
read_knx(file)
}
\arguments{
\item{file}{string. File path to the table extracted from KNX.}
}
\value{

}
\description{
Maps the file name and assigns its respective function
}
