% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFtools.R
\docType{package}
\name{SFtools-package}
\alias{SFtools}
\alias{SFtools-package}
\title{SFtools: Space Filling Based Tools for Data Mining}
\description{
Contains space filling based tools for
machine learning and data mining.
}
\references{
M. Laib and M. Kanevski (2017). Unsupervised Feature Selection Based on Space
Filling Concept, \href{https://arxiv.org/abs/1706.08894}{arXiv:1706.08894}.

J. A. Royle, D. Nychka, An algorithm for the construction of spatial
coverage designs with implementation in Splus, Computers and Geosciences
24 (1997) p. 479–488.

J. Franco, Planification d’expériences numériques en phase exploratoire
pour la simulation des phénomènes complexes, Thesis (2008) 282.

D. Dupuy, C. Helbert, J. Franco (2015). DiceDesign
and DiceEval: Two R Packages for Design and Analysis of Computer
Experiments. Journal of Statistical Software, 65(11), 1-38.
\href{http://www.jstatsoft.org/v65/i11/}{Jstatsoft}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sites.google.com/site/mohamedlaibwebpage/}
  \item Report bugs at \url{https://github.com/mlaib/SFtools/issues}
}

}
\author{
Mohamed Laib \email{Mohamed.Laib@unil.ch}  and

 Mikhail Kanevski \email{Mikhail.Kanevski@unil.ch},

Maintainer: Mohamed Laib \email{laib.med@gmail.com}
}
