% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_twelve.R
\name{sum_twelve}
\alias{sum_twelve}
\title{Sums the last 12 observations}
\usage{
sum_twelve(...)
}
\arguments{
\item{series}{numeric of ts object}
}
\value{
numeric
}
\description{
Sums the last 12 observations
}
\examples{
\dontrun{
sum_twelve(AirPassengers)
}
}
\author{
Obryan Poyser
}
