% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscut.R
\name{tscut}
\alias{tscut}
\title{Suggested starting date}
\usage{
tscut(time_series, freq = 12)
}
\arguments{
\item{time_series}{ts or numeric}

\item{freq}{numeric. Set the frequency if the input is numeric, 12 by default.}
}
\value{
ts object or numeric indexed as year 1
}
\description{
This function uses changing points in variance and mean to suggest a starting date.
}
\examples{
\dontrun{
tscut(AirPassengers)
}
}
